pyproject.toml
[tool.poetry]
name = "pyquizhub"
version = "0.1.0"
description = ""
authors = ["ivan <imtambovtcev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
pyyaml = "^6.0.2"
pytest = "^8.3.4"
requests = "^2.32.3"
httpx = "^0.28.1"

[tool.poetry.scripts]
pyquizhub-setup = "pyquizhub.config.initialize:initialize_system"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"